LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/after-transition-basic-expected.txt


FAIL scroll: after-transition should scroll when back completes assert_equals: expected 0 but got 431
